+++ spi.c
@@ -48,7 +48,7 @@ void SPIInit(void)
 }
 
 // function to write a byte to the SPI port
-void SPIDataWrite(uint32_t ui32DataTx)
+void SPIDataWrite(uint8_t ui32DataTx)
 {
 	ROM_SSIDataPut(SPI_BASE, ui32DataTx);
 }
